third_party/blink/web_tests/fast/css/invalidation/dynamic-selector-list-pseudo-expected.txt

Change classes affecting :-webkit-any and :not.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(t1, null).backgroundColor is transparent
PASS getComputedStyle(t1, null).backgroundColor is green
PASS getComputedStyle(t2, null).backgroundColor is transparent
PASS getComputedStyle(t2, null).backgroundColor is green
PASS getComputedStyle(t3, null).backgroundColor is transparent
PASS getComputedStyle(t3, null).backgroundColor is green
PASS getComputedStyle(t4, null).backgroundColor is transparent
PASS getComputedStyle(t4, null).backgroundColor is green
PASS getComputedStyle(t5, null).backgroundColor is transparent
PASS getComputedStyle(t5, null).backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE

Background should be green
Background should be green
Background should be green
Background should be green
Background should be green
